// src/types/sensor.ts
export interface GaugeThreshold {
  warn: number;
  crit: number;
  inverted?: boolean;
}

export type GaugeThresholds = Record<string, GaugeThreshold>;